% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_de_dnn.R
\name{de.dnn}
\alias{de.dnn}
\title{Get a de-DNN estimate using R implementation}
\usage{
de.dnn(X, Y, X.test, s.size = 2, bc.p = 2)
}
\arguments{
\item{X}{Matrix of covariates}

\item{Y}{Matrix of responses}

\item{X.test}{Matrix of a single test observation for which we want to get an estimate}

\item{s.size}{Size of subsample used in estimation}

\item{bc.p}{Ratio between the subsampling scales: bc.p = s_2/s_1}
}
\description{
Get a de-DNN estimate using R implementation
}
