% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_de_dnn.R
\name{de.dnn_no_dist}
\alias{de.dnn_no_dist}
\title{Get a de-DNN estimate using R implementation
This version of the algorithm assumes that the Euclidean distance between
the train and test observations has already been calculated and used to order the responses}
\usage{
de.dnn_no_dist(ordered_Y, n, p, s.size = 2, bc.p = 2)
}
\arguments{
\item{ordered_Y}{Matrix of Y values ordered by euclidean distance}

\item{n}{Number of rows in the training data}

\item{p}{Number of columns in the training data}

\item{s.size}{Size of subsample used in estimation}

\item{bc.p}{Ratio between the subsampling scales: bc.p = s_2/s_1}
}
\description{
Get a de-DNN estimate using R implementation
This version of the algorithm assumes that the Euclidean distance between
the train and test observations has already been calculated and used to order the responses
}
