% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{est_effect}
\alias{est_effect}
\title{Estimate a treatment effect using the two-scale DNN estimator}
\usage{
est_effect(X, W, Y, X_test, W_0, t = 50, estimate_var = F, ...)
}
\arguments{
\item{X}{Matrix of covariates}

\item{W}{Matrix of treatment assignments}

\item{Y}{Matrix of responses}

\item{X_test}{Matrix of test observations}

\item{W_0}{Optional boolean feature screening vector}

\item{t}{Max size of tuning sequence. Default is 50}

\item{estimate_var}{Boolean for estimating variance using bootstrap}

\item{...}{Extra arguments to be passed to boot when estimating variance (e.g. ncpus, parallel, R)}
}
\description{
Estimate a treatment effect using the two-scale DNN estimator
}
