% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{tdnn}
\alias{tdnn}
\title{Estimate a treatment effect using the two-scale DNN estimator}
\usage{
tdnn(X, Y, X_test, W_0 = NULL, t = 50)
}
\arguments{
\item{X}{matrix of covariates}

\item{Y}{matrix of responses}

\item{X_test}{matrix of test observations}

\item{W_0}{optional boolean feature screening vector}

\item{t}{max size of tuning sequence. Default is 50}
}
\description{
Estimate a treatment effect using the two-scale DNN estimator
}
