% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdnn_reg_cv.R
\name{tdnn_reg_cv}
\alias{tdnn_reg_cv}
\title{Get a tuned tdnn estimate}
\usage{
tdnn_reg_cv(
  X,
  Y,
  X_test,
  tuning_iter = 200,
  W_0 = NULL,
  s_seq = NULL,
  c_seq = NULL,
  n_prop = 0.5,
  estimate_variance = F,
  bootstrap_iter = 1000,
  n_threads = NULL,
  verbose = F
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{Y}{Matrix of responses}

\item{X_test}{Matrix of test observations for which we want to get estimates}

\item{tuning_iter}{Number of iterations for MC resampling when tuning parameters. Default is 200}

\item{W_0}{Optional integer vector with 1 corresponding to columns that should be used for estimation. Default value is NULL}

\item{s_seq}{Sequence of potential values for \eqn{s_1}. If a sequence is not provided, the default is \code{seq(1,50,1)}}

\item{c_seq}{Sequence of potential values for \eqn{c}. If a sequence is not provided, the default is \code{seq(1.25, 3, 0.25)}}

\item{n_prop}{If \eqn{s_{2} > n\dot \text{n_prop}}, default to using 1-NN estimate.}

\item{estimate_variance}{Boolean flag for estimating variance. Default is FALSE.}

\item{bootstrap_iter}{Number of bootstrap iterations to use when estimating variance. Default is 1000.}

\item{n_threads}{Number of threads to use when calculating bootstrap variance. Default is to use all available threads.}

\item{verbose}{Print which step the method is currently calculating in the console.}
}
\description{
Get a tuned tdnn estimate
}
