% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdnn_reg_tune.R
\name{tdnn_reg_tune}
\alias{tdnn_reg_tune}
\title{Get a tuned tdnn estimate}
\usage{
tdnn_reg_tune(
  X,
  Y,
  X_test,
  W_0 = NULL,
  c_val = 2,
  B_NN = 20,
  n_prop = 0.5,
  estimate_variance = F,
  bootstrap_iter = 1000,
  n_threads = NULL,
  verbose = F
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{Y}{Matrix of responses}

\item{X_test}{Matrix of test observations for which we want to get estimates}

\item{W_0}{Optional integer vector with 1 corresponding to columns that should be used for estimation. Default value is NULL}

\item{c_val}{Sequence of potential values for \eqn{c}. If a sequence is not provided, the default is 2.}

\item{B_NN}{Number of nearest neighbors to use during tuning. Default is 20.}

\item{n_prop}{If \eqn{s_{2} > n\dot \text{n_prop}}, default to using 1-NN estimate.}

\item{estimate_variance}{Boolean flag for estimating variance. Default is FALSE.}

\item{bootstrap_iter}{Number of bootstrap iterations to use when estimating variance. Default is 1000.}

\item{n_threads}{Number of threads to use when calculating bootstrap variance. Default is to use all available threads.}

\item{verbose}{Print which step the method is currently calculating in the console as well as debug information.}
}
\description{
Get a tuned tdnn estimate
}
