% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_de_dnn.R
\name{tune_de_dnn_no_dist_test_mat}
\alias{tune_de_dnn_no_dist_test_mat}
\title{function that allows for multiple test observations}
\usage{
tune_de_dnn_no_dist_test_mat(
  X,
  Y,
  X_test,
  c = 2,
  B_NN = 20,
  scale_p = 1,
  debug = F
)
}
\description{
function that allows for multiple test observations
}
