% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_de_dnn.R
\name{tune_s}
\alias{tune_s}
\title{Tune the subsampling scale s using the R implementation}
\usage{
tune_s(X, Y, X_test, s_seq, c)
}
\arguments{
\item{X}{Matrix of covariates}

\item{Y}{Matrix of responses}

\item{X_test}{Matrix of a single test observation for which we want to get an estimate}

\item{s_seq}{Vector of s values to use for tuning}

\item{c}{Ratio between the subsampling scales: c = s_2/s_1
#' @export}
}
\description{
Tune the subsampling scale s using the R implementation
}
