% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_de_dnn.R
\name{tune_s_no_dist}
\alias{tune_s_no_dist}
\title{Tune the subsampling scale s using the R implementation
This version of the algorithm assumes that the Euclidean distance between the
train and test observations has already been calculated and used to order the responses}
\usage{
tune_s_no_dist(ordered_Y, n, p, s_seq, c)
}
\arguments{
\item{ordered_Y}{Matrix of Y values ordered by euclidean distance}

\item{n}{Number of rows in the training data}

\item{p}{Number of columns in the training data}

\item{s_seq}{Vector of s values to use for tuning}

\item{c}{Ratio between the subsampling scales: c = s_2/s_1
#' @export}
}
\description{
Tune the subsampling scale s using the R implementation
This version of the algorithm assumes that the Euclidean distance between the
train and test observations has already been calculated and used to order the responses
}
