% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{est_reg_fn}
\alias{est_reg_fn}
\title{Estimate a regression function using the two-scale DNN estimator}
\usage{
est_reg_fn(X, Y, X_test, W_0 = NULL, t = 50)
}
\arguments{
\item{X}{matrix of covariates}

\item{Y}{matrix of responses}

\item{X_test}{matrix of test observations}

\item{W_0}{optional Boolean feature screening vector}

\item{t}{max size of tuning sequence. Default is 50}
}
\description{
Estimate a regression function using the two-scale DNN estimator
}
