% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdnn_reg.R
\name{tdnn_reg}
\alias{tdnn_reg}
\title{Get tdnn estimate}
\usage{
tdnn_reg(
  X,
  Y,
  X_test,
  s_1,
  c,
  W_0 = NULL,
  n_prop = 0.5,
  estimate_variance = F,
  bootstrap_iter = 1000,
  n_threads = NULL,
  verbose = F
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{Y}{Matrix of responses}

\item{X_test}{Matrix of test observations for which we want to get estimates}

\item{s_1}{Value of s_1.}

\item{c}{Value of c.}

\item{W_0}{Optional integer vector with 1 corresponding to columns that should be used for estimation. Default value is NULL}

\item{n_prop}{If \eqn{s_{2} > n\dot \text{n_prop}}, default to using 1-NN estimate.}

\item{estimate_variance}{Boolean flag for estimating variance. Default is FALSE.}

\item{bootstrap_iter}{Number of bootstrap iterations to use when estimating variance. Default is 1000.}

\item{n_threads}{Number of threads to use when calculating bootstrap variance. Default is to use all available threads.}

\item{verbose}{Print which step the method is currently calculating in the console.}
}
\description{
Get tdnn estimate
}
